requirements.txt
python-dotenv
langchain
langchain-community
langchain-openai
openai
gradio>=4.0.0
SQLAlchemy